+++ CHANGES.rst
@@ -6,9 +6,13 @@ Changes
 1.0 (unreleased)
 ================
 
+- Add support for Python 3.9, 3.10, 3.11.
+
 - Drop support for Python 2.7, 3.5, 3.6.
 
 - Use GitHub actions as CI.
+
+- Update tests to support ``zope.testing >= 5``.
 
 
 0.1 (2020-08-07)